Parser for union types
This also simplifies the parsers for all types by removing all the
Enum.join which were scattered everywhere.

Also removed the conditional treatment of joining atoms in the type parser by
making it a string. This removes the need for putting the atom parser in
Fika.Parser.Common, so it's moved to Fika.Parser.LiteralExps where it belongs.